// types/gql/graphql.ts
export type InputMaybe<T> = T | null | undefined;
// ...
export type CharacterConnection = {
	__typename?: "CharacterConnection";
	edges?: Maybe<Array<Maybe<CharacterEdge>>>;
	nodes?: Maybe<Array<Maybe<Character>>>;
	/** The pagination information */
	pageInfo?: Maybe<PageInfo>;
};
// ...
export type FuzzyDate = {
	__typename?: "FuzzyDate";
	/** Numeric Day (24) */
	day?: Maybe<Scalars["Int"]["output"]>;
	/** Numeric Month (3) */
	month?: Maybe<Scalars["Int"]["output"]>;
	/** Numeric Year (2017) */
	year?: Maybe<Scalars["Int"]["output"]>;
};
// ...
export type MediaCoverImage = {
	__typename?: "MediaCoverImage";
	/** Average #hex color of cover image */
	color?: Maybe<Scalars["String"]["output"]>;
	/** The cover image url of the media at its largest size. If this size isn't available, large will be provided instead. */
	extraLarge?: Maybe<Scalars["String"]["output"]>;
	/** The cover image url of the media at a large size */
	large?: Maybe<Scalars["String"]["output"]>;
	/** The cover image url of the media at medium size */
	medium?: Maybe<Scalars["String"]["output"]>;
};
// ...
export type MediaTitle = {
	__typename?: "MediaTitle";
	/** The official english title */
	english?: Maybe<Scalars["String"]["output"]>;
	/** Official title in it's native language */
	native?: Maybe<Scalars["String"]["output"]>;
	/** The romanization of the native language title */
	romaji?: Maybe<Scalars["String"]["output"]>;
	/** The currently authenticated users preferred title language. Default romaji for non-authenticated */
	userPreferred?: Maybe<Scalars["String"]["output"]>;
};
// ...
export type Studio = {
	__typename?: "Studio";
	/** The amount of user's who have favourited the studio */
	favourites?: Maybe<Scalars["Int"]["output"]>;
	/** The id of the studio */
	id: Scalars["Int"]["output"];
	/** If the studio is an animation studio or a different kind of company */
	isAnimationStudio: Scalars["Boolean"]["output"];
	/** If the studio is marked as favourite by the currently authenticated user */
	isFavourite: Scalars["Boolean"]["output"];
	/** The media the studio has worked on */
	media?: Maybe<MediaConnection>;
	/** The name of the studio */
	name: Scalars["String"]["output"];
	/** The url for the studio page on the AniList website */
	siteUrl?: Maybe<Scalars["String"]["output"]>;
};

/** Animation or production company */
export type StudioMediaArgs = {
	isMain?: InputMaybe<Scalars["Boolean"]["input"]>;
	onList?: InputMaybe<Scalars["Boolean"]["input"]>;
	page?: InputMaybe<Scalars["Int"]["input"]>;
	perPage?: InputMaybe<Scalars["Int"]["input"]>;
	sort?: InputMaybe<Array<InputMaybe<MediaSort>>>;
};

export type StudioConnection = {
	__typename?: "StudioConnection";
	edges?: Maybe<Array<Maybe<StudioEdge>>>;
	nodes?: Maybe<Array<Maybe<Studio>>>;
	/** The pagination information */
	pageInfo?: Maybe<PageInfo>;
};
// ...
export type AnimeDetailsDataFragment = {
	__typename?: "Media";
	id: number;
	popularity?: number | null;
	averageScore?: number | null;
	description?: string | null;
	studios?: {
		__typename?: "StudioConnection";
		nodes?: Array<{ __typename?: "Studio"; name: string } | null> | null;
	} | null;
	coverImage?: {
		__typename?: "MediaCoverImage";
		extraLarge?: string | null;
	} | null;
	startDate?: { __typename?: "FuzzyDate"; year?: number | null } | null;
	title?: { __typename?: "MediaTitle"; userPreferred?: string | null } | null;
	characters?: {
		__typename?: "CharacterConnection";
		nodes?: Array<{
			__typename?: "Character";
			id: number;
			gender?: string | null;
			name?: {
				__typename?: "CharacterName";
				userPreferred?: string | null;
			} | null;
			image?: { __typename?: "CharacterImage"; medium?: string | null } | null;
		} | null> | null;
	} | null;
} & { " $fragmentName"?: "AnimeDetailsDataFragment" };

export type GetMediaQueryVariables = Exact<{
	id?: InputMaybe<Scalars["Int"]["input"]>;
}>;

export type GetMediaQuery = {
	__typename?: "Query";
	Media?:
		| ({ __typename?: "Media" } & {
				" $fragmentRefs"?: {
					AnimeDetailsDataFragment: AnimeDetailsDataFragment;
				};
		  })
		| null;
};
// ...
export const AnimeDetailsDataFragmentDoc = {
	kind: "Document",
	definitions: [
		{
			kind: "FragmentDefinition",
			name: { kind: "Name", value: "AnimeDetailsData" },
			typeCondition: {
				kind: "NamedType",
				name: { kind: "Name", value: "Media" },
			},
			selectionSet: {
				kind: "SelectionSet",
				selections: [
					{ kind: "Field", name: { kind: "Name", value: "id" } },
					{
						kind: "Field",
						name: { kind: "Name", value: "studios" },
						arguments: [
							{
								kind: "Argument",
								name: { kind: "Name", value: "isMain" },
								value: { kind: "BooleanValue", value: true },
							},
						],
						selectionSet: {
							kind: "SelectionSet",
							selections: [
								{
									kind: "Field",
									name: { kind: "Name", value: "nodes" },
									selectionSet: {
										kind: "SelectionSet",
										selections: [
											{ kind: "Field", name: { kind: "Name", value: "name" } },
										],
									},
								},
							],
						},
					},
					{
						kind: "Field",
						name: { kind: "Name", value: "coverImage" },
						selectionSet: {
							kind: "SelectionSet",
							selections: [
								{ kind: "Field", name: { kind: "Name", value: "extraLarge" } },
							],
						},
					},
					{
						kind: "Field",
						name: { kind: "Name", value: "startDate" },
						selectionSet: {
							kind: "SelectionSet",
							selections: [
								{ kind: "Field", name: { kind: "Name", value: "year" } },
							],
						},
					},
					{ kind: "Field", name: { kind: "Name", value: "popularity" } },
					{ kind: "Field", name: { kind: "Name", value: "averageScore" } },
					{
						kind: "Field",
						name: { kind: "Name", value: "description" },
						arguments: [
							{
								kind: "Argument",
								name: { kind: "Name", value: "asHtml" },
								value: { kind: "BooleanValue", value: true },
							},
						],
					},
					{
						kind: "Field",
						name: { kind: "Name", value: "title" },
						selectionSet: {
							kind: "SelectionSet",
							selections: [
								{
									kind: "Field",
									name: { kind: "Name", value: "userPreferred" },
								},
							],
						},
					},
					{
						kind: "Field",
						name: { kind: "Name", value: "characters" },
						arguments: [
							{
								kind: "Argument",
								name: { kind: "Name", value: "role" },
								value: { kind: "EnumValue", value: "MAIN" },
							},
						],
						selectionSet: {
							kind: "SelectionSet",
							selections: [
								{
									kind: "Field",
									name: { kind: "Name", value: "nodes" },
									selectionSet: {
										kind: "SelectionSet",
										selections: [
											{ kind: "Field", name: { kind: "Name", value: "id" } },
											{
												kind: "Field",
												name: { kind: "Name", value: "name" },
												selectionSet: {
													kind: "SelectionSet",
													selections: [
														{
															kind: "Field",
															name: { kind: "Name", value: "userPreferred" },
														},
													],
												},
											},
											{
												kind: "Field",
												name: { kind: "Name", value: "gender" },
											},
											{
												kind: "Field",
												name: { kind: "Name", value: "image" },
												selectionSet: {
													kind: "SelectionSet",
													selections: [
														{
															kind: "Field",
															name: { kind: "Name", value: "medium" },
														},
													],
												},
											},
										],
									},
								},
							],
						},
					},
				],
			},
		},
	],
} as unknown as DocumentNode<AnimeDetailsDataFragment, unknown>;
// ...
export const GetMediaDocument = {
	kind: "Document",
	definitions: [
		{
			kind: "OperationDefinition",
			operation: "query",
			name: { kind: "Name", value: "GetMedia" },
			variableDefinitions: [
				{
					kind: "VariableDefinition",
					variable: { kind: "Variable", name: { kind: "Name", value: "id" } },
					type: { kind: "NamedType", name: { kind: "Name", value: "Int" } },
				},
			],
			selectionSet: {
				kind: "SelectionSet",
				selections: [
					{
						kind: "Field",
						name: { kind: "Name", value: "Media" },
						arguments: [
							{
								kind: "Argument",
								name: { kind: "Name", value: "id" },
								value: {
									kind: "Variable",
									name: { kind: "Name", value: "id" },
								},
							},
						],
						selectionSet: {
							kind: "SelectionSet",
							selections: [
								{
									kind: "FragmentSpread",
									name: { kind: "Name", value: "AnimeDetailsData" },
								},
							],
						},
					},
				],
			},
		},
		{
			kind: "FragmentDefinition",
			name: { kind: "Name", value: "AnimeDetailsData" },
			typeCondition: {
				kind: "NamedType",
				name: { kind: "Name", value: "Media" },
			},
			selectionSet: {
				kind: "SelectionSet",
				selections: [
					{ kind: "Field", name: { kind: "Name", value: "id" } },
					{
						kind: "Field",
						name: { kind: "Name", value: "studios" },
						arguments: [
							{
								kind: "Argument",
								name: { kind: "Name", value: "isMain" },
								value: { kind: "BooleanValue", value: true },
							},
						],
						selectionSet: {
							kind: "SelectionSet",
							selections: [
								{
									kind: "Field",
									name: { kind: "Name", value: "nodes" },
									selectionSet: {
										kind: "SelectionSet",
										selections: [
											{ kind: "Field", name: { kind: "Name", value: "name" } },
										],
									},
								},
							],
						},
					},
					{
						kind: "Field",
						name: { kind: "Name", value: "coverImage" },
						selectionSet: {
							kind: "SelectionSet",
							selections: [
								{ kind: "Field", name: { kind: "Name", value: "extraLarge" } },
							],
						},
					},
					{
						kind: "Field",
						name: { kind: "Name", value: "startDate" },
						selectionSet: {
							kind: "SelectionSet",
							selections: [
								{ kind: "Field", name: { kind: "Name", value: "year" } },
							],
						},
					},
					{ kind: "Field", name: { kind: "Name", value: "popularity" } },
					{ kind: "Field", name: { kind: "Name", value: "averageScore" } },
					{
						kind: "Field",
						name: { kind: "Name", value: "description" },
						arguments: [
							{
								kind: "Argument",
								name: { kind: "Name", value: "asHtml" },
								value: { kind: "BooleanValue", value: true },
							},
						],
					},
					{
						kind: "Field",
						name: { kind: "Name", value: "title" },
						selectionSet: {
							kind: "SelectionSet",
							selections: [
								{
									kind: "Field",
									name: { kind: "Name", value: "userPreferred" },
								},
							],
						},
					},
					{
						kind: "Field",
						name: { kind: "Name", value: "characters" },
						arguments: [
							{
								kind: "Argument",
								name: { kind: "Name", value: "role" },
								value: { kind: "EnumValue", value: "MAIN" },
							},
						],
						selectionSet: {
							kind: "SelectionSet",
							selections: [
								{
									kind: "Field",
									name: { kind: "Name", value: "nodes" },
									selectionSet: {
										kind: "SelectionSet",
										selections: [
											{ kind: "Field", name: { kind: "Name", value: "id" } },
											{
												kind: "Field",
												name: { kind: "Name", value: "name" },
												selectionSet: {
													kind: "SelectionSet",
													selections: [
														{
															kind: "Field",
															name: { kind: "Name", value: "userPreferred" },
														},
													],
												},
											},
											{
												kind: "Field",
												name: { kind: "Name", value: "gender" },
											},
											{
												kind: "Field",
												name: { kind: "Name", value: "image" },
												selectionSet: {
													kind: "SelectionSet",
													selections: [
														{
															kind: "Field",
															name: { kind: "Name", value: "medium" },
														},
													],
												},
											},
										],
									},
								},
							],
						},
					},
				],
			},
		},
	],
} as unknown as DocumentNode<GetMediaQuery, GetMediaQueryVariables>;